+++ src/Context.php
@@ -21,17 +21,15 @@ interface Context
      * In case the rounding mode is irrelevant, for example in AutoContext, this method MUST throw an exception if a
      * rounding mode other than RoundingMode::UNNECESSARY is used.
      *
-     * @psalm-param RoundingMode::* $roundingMode
-     *
-     * @param BigNumber $amount       The amount.
-     * @param Currency  $currency     The target currency.
-     * @param int       $roundingMode The rounding mode.
+     * @param BigNumber    $amount       The amount.
+     * @param Currency     $currency     The target currency.
+     * @param RoundingMode $roundingMode The rounding mode.
      *
      * @return BigDecimal
      *
      * @throws RoundingNecessaryException If the result cannot be represented at the required scale without rounding.
      */
-    public function applyTo(BigNumber $amount, Currency $currency, int $roundingMode) : BigDecimal;
+    public function applyTo(BigNumber $amount, Currency $currency, RoundingMode $roundingMode) : BigDecimal;
 
     /**
      * Returns the step used by this context.