+++ src/Context/CustomContext.php
@@ -5,10 +5,10 @@ declare(strict_types=1);
 namespace Brick\Money\Context;
 
 use Brick\Math\BigDecimal;
+use Brick\Math\BigNumber;
+use Brick\Math\RoundingMode;
 use Brick\Money\Context;
 use Brick\Money\Currency;
-
-use Brick\Math\BigNumber;
 
 /**
  * Adjusts a number to a custom scale, and optionally step.
@@ -40,7 +40,7 @@ final class CustomContext implements Context
     /**
      * {@inheritdoc}
      */
-    public function applyTo(BigNumber $amount, Currency $currency, int $roundingMode) : BigDecimal
+    public function applyTo(BigNumber $amount, Currency $currency, RoundingMode $roundingMode) : BigDecimal
     {
         if ($this->step === 1) {
             return $amount->toScale($this->scale, $roundingMode);