+++ src/Context/DefaultContext.php
@@ -9,6 +9,7 @@ use Brick\Money\Currency;
 
 use Brick\Math\BigDecimal;
 use Brick\Math\BigNumber;
+use Brick\Math\RoundingMode;
 
 /**
  * Adjusts a number to the default scale for the currency.
@@ -18,7 +19,7 @@ final class DefaultContext implements Context
     /**
      * @inheritdoc
      */
-    public function applyTo(BigNumber $amount, Currency $currency, int $roundingMode) : BigDecimal
+    public function applyTo(BigNumber $amount, Currency $currency, RoundingMode $roundingMode) : BigDecimal
     {
         return $amount->toScale($currency->getDefaultFractionDigits(), $roundingMode);
     }