+++ tests/Context/CashContextTest.php
@@ -20,7 +20,7 @@ class CashContextTest extends AbstractTestCase
     /**
      * @dataProvider providerApplyTo
      */
-    public function testApplyTo(int $step, string $amount, string $currency, int $roundingMode, string $expected) : void
+    public function testApplyTo(int $step, string $amount, string $currency, RoundingMode $roundingMode, string $expected) : void
     {
         $amount = BigNumber::of($amount);
         $currency = Currency::of($currency);